refactor(search): rename components and relocate them inside the project
LIIKUNTA-425
