+++ apps/sports-helsinki/src/domain/search/combinedSearch/search.module.scss
